+++ java/src/com/presnakov/hotelBookingSystem/servlet/RoomOrderServlet.java
@@ -17,7 +17,7 @@ public class RoomOrderServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        var roomId = Long.valueOf(req.getParameter("roomId"));
+        var roomId = Integer.valueOf(req.getParameter("roomId"));
         req.setAttribute("orders", roomOrderService.findAllByRoomId(roomId));
         req.getRequestDispatcher(JspHelper.getPath("orders"))
                 .forward(req, resp);
